作り直すためにプッシュ

+++ TrainingGraphicsAPI/Object.cpp
@@ -1,7 +1,6 @@
 #include "Object.h"
 
-
-bool Object::ObjectCreate(ID3D11Device* Device, Vertex* VertexList, int VertexNum, unsigned short* IndexList, int IndexNum, ID3D11DeviceContext* Context)
+bool Object::ObjectCreate(ID3D11Device* Device, Vertex* VertexList, int VertexNum, unsigned long* IndexList, int IndexNum, ID3D11DeviceContext* Context)
 {
 	auto VertexByteSize = sizeof(Vertex) * VertexNum;
 	auto IndexByteSize = sizeof(unsigned short) * IndexNum;
@@ -59,7 +58,7 @@ bool Object::ObjectCreate(ID3D11Device* Device, Vertex* VertexList, int VertexNu
 	return true;
 }
 
-bool Object::ObjectCreate(ID3D12Device* Device, Vertex* VertexList, int VertexNum, unsigned int* IndexList, int IndexNum)
+bool Object::ObjectCreate(ID3D12Device* Device, Vertex* VertexList, int VertexNum, unsigned long* IndexList, int IndexNum)
 {
 	auto VertexByteSize = sizeof(Vertex) * VertexNum;
 	auto IndexByteSize = sizeof(unsigned int) * IndexNum;
@@ -545,4 +544,3 @@ void Object::Update(ID3D11DeviceContext* Context)
 	Context->UpdateSubresource(m_11ConstantBuffer.Get(), 0, NULL, &cb, 0, 0);
 
 }
-
